+++ client/src/shared/ui/card.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../lib/index";
+import { cn } from "../lib";
 
 function Card({ className, ...props }: React.ComponentProps<"div">) {
   return (
