+++ client/src/shared/ui/input.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../lib/index";
+import { cn } from "../lib";
 
 function Input({ className, type, ...props }: React.ComponentProps<"input">) {
   return (
