+++ client/src/shared/ui/table.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../lib/index";
+import { cn } from "../lib";
 
 function Table({ className, ...props }: React.ComponentProps<"table">) {
   return (
